feat(blockchain): enhance BlockHeader timestamp type and implement genesis block creation

+++ crates/btc-core/src/block/header.rs
@@ -9,7 +9,7 @@ pub struct BlockHeader {
     pub version: u32,
     pub previous_block_hash: BlockHash,
     pub merkle_root: MerkleRoot,
-    pub timestamp: u32,
+    pub timestamp: u64,
     pub bits: u32,
     pub nonce: u32,
 }
